+++ sys/sys.c
@@ -25,15 +25,15 @@ extern char *sys$getenv(const char *key) {
     return getenv(key);
 }
 
-extern int sys$memcmp(const void *dst, const void *src, size_t n) {
+extern int sys$memcmp(const void *dst, const void *src, u64 n) {
     return memcmp(dst, src, n);
 }
 
-extern void *sys$memcpy(void *dst, const void *src, size_t n) {
+extern void *sys$memcpy(void *dst, const void *src, u64 n) {
     return memcpy(dst, src, n);
 }
 
-extern void *sys$memset(void *dst, int c, size_t n) {
+extern void *sys$memset(void *dst, int c, u64 n) {
     return memset(dst, c, n);
 }
 
@@ -41,15 +41,15 @@ extern char *sys$strdup(const char *s) {
     return strdup(s);
 }
 
-extern size_t sys$strlcpy(char *dst, const char *src, size_t n) {
+extern u64 sys$strlcpy(char *dst, const char *src, u64 n) {
     return strlcpy(dst, src, n);
 }
 
-extern size_t sys$strlen(const char *s) {
+extern u64 sys$strlen(const char *s) {
     return strlen(s);
 }
 
-extern void *sys$malloc(size_t n) {
+extern void *sys$malloc(u64 n) {
     return malloc(n);
 }
 
@@ -57,7 +57,7 @@ extern void sys$free(void *ptr) {
     free(ptr);
 }
 
-extern void *sys$realloc(void *ptr, size_t n) {
+extern void *sys$realloc(void *ptr, u64 n) {
     return realloc(ptr, n);
 }
